Downgrade to use UInt

+++ VoiceCraft.Client/VoiceCraft.Client/Audio/Effects/ClientVisibilityEffect.cs
@@ -6,7 +6,7 @@ namespace VoiceCraft.Client.Audio.Effects;
 
 public class ClientVisibilityEffect : VisibilityEffect
 {
-    public override void Process(VoiceCraftEntity from, VoiceCraftEntity to, ulong effectBitmask, Span<float> data, int count)
+    public override void Process(VoiceCraftEntity from, VoiceCraftEntity to, uint effectBitmask, Span<float> data, int count)
     {
     }
 }
